+++ docs/changelog.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* :feature:`327` Update container class to have contextual custom properties
+
 * :release:`7.10.1 <2021-10-27>`
 * :feature:`323` Add seal ["ultra-clear", "foil"] to FLAT384WHITETC
 * :support:`318` Update all pre-commit hooks except pylint to latest versions